收费课时做完

+++ TraininginsTitutions/src/cn/huizhi/pojo/Student.java
@@ -19,13 +19,13 @@ public class Student {
 	private String theSurrogate;
 	private String remarks;
 	private Integer feeCategory;
-	private double studentHour;
+	private Integer studentHour;
 	private ChildrenesClassStudnet childrenesClassStudnet;
 
 	public Student(Integer studentId, String studentName, Integer studentSex, String studentIDCard, String studentBirth,
 			String school, String parentName, String telephone, String familyAddress, String registrationChannels,
 			String registrationConsultant, String startTime, String endTime, String theSurrogate, String remarks,
-			Integer feeCategory, double studentHour, ChildrenesClassStudnet childrenesClassStudnet) {
+			Integer feeCategory, Integer studentHour, ChildrenesClassStudnet childrenesClassStudnet) {
 		super();
 		this.studentId = studentId;
 		this.studentName = studentName;
@@ -67,11 +67,11 @@ public class Student {
 		this.studentId = studentId;
 	}
 
-	public double getStudentHour() {
+	public Integer getStudentHour() {
 		return studentHour;
 	}
 
-	public void setStudentHour(double studentHour) {
+	public void setStudentHour(Integer studentHour) {
 		this.studentHour = studentHour;
 	}
